Simulate from Cox MSM Model

Obtain samples from a specified Cox Marginal Structural Model using the frugal parameterization.

msm_samp(
  n,
  dat = NULL,
  qtls = NULL,
  T,
  formulas,
  family,
  pars,
  link = NULL,
  method = "inversion",
  control = list()
)

coxSamp(n, T, formulas, family, pars, link = NULL, control = list())

cox_samp(n, T, formulas, family, pars, link = NULL, control = list())

Arguments

n

number of samples

dat

optional data frame for plasmode simulation

qtls

associated quantiles from dat for plasmode simulation

T

number of time points

formulas

list of formulas to use

family

list of families to use

pars

list of parameter settings to use

link

link functions for GLM-like models

method

sampling method (defaults to "inversion")

control

list of control parameters

Value

An object of class survivl_dat containing the simulated data.

Details

Samples from a Marginal Structural Model specified by a frugal parameterization; that is, one must specify the marginal model (i.e. the dependence of the outcome on the treatments), the distribution of the covariates, observed confounders and treatments, and then a copula to join the distribution of the outcome to that of the confounders.

Among the left-hand sides of outcome variables, the variable 'Cen' has a special meaning as censoring. This keyword can be changed to something else by using the argument censor in the control list.

Functions

coxSamp(): old name

cox_samp(): old name